Reorganised the code
